LayoutTests/imported/w3c/web-platform-tests/trusted-types/should-trusted-type-policy-creation-be-blocked-by-csp-003-expected.txt

CONSOLE MESSAGE: Refused to connect to ws://common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: The Content Security Policy 'trusted-types A' was delivered in report-only mode, but does not specify a 'report-to'; the policy will have no effect. Please either add a 'report-to' directive, or deliver the policy via the 'Content-Security-Policy' header.
CONSOLE MESSAGE: The Content Security Policy ' trusted-types B' was delivered in report-only mode, but does not specify a 'report-to'; the policy will have no effect. Please either add a 'report-to' directive, or deliver the policy via the 'Content-Security-Policy' header.
CONSOLE MESSAGE: The Content Security Policy ' trusted-types C' was delivered in report-only mode, but does not specify a 'report-to'; the policy will have no effect. Please either add a 'report-to' directive, or deliver the policy via the 'Content-Security-Policy' header.
CONSOLE MESSAGE: The Content Security Policy ' trusted-types tt-policy-name' was delivered in report-only mode, but does not specify a 'report-to'; the policy will have no effect. Please either add a 'report-to' directive, or deliver the policy via the 'Content-Security-Policy' header.
CONSOLE MESSAGE: The Content Security Policy ' trusted-types D' was delivered in report-only mode, but does not specify a 'report-to'; the policy will have no effect. Please either add a 'report-to' directive, or deliver the policy via the 'Content-Security-Policy' header.
CONSOLE MESSAGE: The Content Security Policy ' trusted-types E' was delivered in report-only mode, but does not specify a 'report-to'; the policy will have no effect. Please either add a 'report-to' directive, or deliver the policy via the 'Content-Security-Policy' header.
CONSOLE MESSAGE: The Content Security Policy ' trusted-types tt-policy-name' was delivered in report-only mode, but does not specify a 'report-to'; the policy will have no effect. Please either add a 'report-to' directive, or deliver the policy via the 'Content-Security-Policy' header.
CONSOLE MESSAGE: [Report Only] Refused to create a TrustedTypePolicy named 'tt-policy-name' because it violates the following Content Security Policy directive: "trusted-types A"
CONSOLE MESSAGE: [Report Only] Refused to create a TrustedTypePolicy named 'tt-policy-name' because it violates the following Content Security Policy directive: "trusted-types B"
CONSOLE MESSAGE: [Report Only] Refused to create a TrustedTypePolicy named 'tt-policy-name' because it violates the following Content Security Policy directive: "trusted-types C"
CONSOLE MESSAGE: [Report Only] Refused to create a TrustedTypePolicy named 'tt-policy-name' because it violates the following Content Security Policy directive: "trusted-types D"
CONSOLE MESSAGE: [Report Only] Refused to create a TrustedTypePolicy named 'tt-policy-name' because it violates the following Content Security Policy directive: "trusted-types E"
CONSOLE MESSAGE: Refused to connect to ws://common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: The Content Security Policy 'trusted-types A' was delivered in report-only mode, but does not specify a 'report-to'; the policy will have no effect. Please either add a 'report-to' directive, or deliver the policy via the 'Content-Security-Policy' header.
CONSOLE MESSAGE: Refused to create a TrustedTypePolicy named 'tt-policy-name' because it violates the following Content Security Policy directive: "trusted-types B"
CONSOLE MESSAGE: Refused to create a TrustedTypePolicy named 'tt-policy-name' because it violates the following Content Security Policy directive: "trusted-types C"
CONSOLE MESSAGE: Refused to create a TrustedTypePolicy named 'tt-policy-name' because it violates the following Content Security Policy directive: "trusted-types D"
CONSOLE MESSAGE: Refused to create a TrustedTypePolicy named 'tt-policy-name' because it violates the following Content Security Policy directive: "trusted-types E"
CONSOLE MESSAGE: [Report Only] Refused to create a TrustedTypePolicy named 'tt-policy-name' because it violates the following Content Security Policy directive: "trusted-types A"
CONSOLE MESSAGE: Refused to connect to ws://common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: The Content Security Policy 'trusted-types B' was delivered in report-only mode, but does not specify a 'report-to'; the policy will have no effect. Please either add a 'report-to' directive, or deliver the policy via the 'Content-Security-Policy' header.
CONSOLE MESSAGE: The Content Security Policy ' trusted-types C' was delivered in report-only mode, but does not specify a 'report-to'; the policy will have no effect. Please either add a 'report-to' directive, or deliver the policy via the 'Content-Security-Policy' header.
CONSOLE MESSAGE: The Content Security Policy ' trusted-types tt-policy-name' was delivered in report-only mode, but does not specify a 'report-to'; the policy will have no effect. Please either add a 'report-to' directive, or deliver the policy via the 'Content-Security-Policy' header.
CONSOLE MESSAGE: The Content Security Policy ' trusted-types D' was delivered in report-only mode, but does not specify a 'report-to'; the policy will have no effect. Please either add a 'report-to' directive, or deliver the policy via the 'Content-Security-Policy' header.
CONSOLE MESSAGE: The Content Security Policy ' trusted-types E' was delivered in report-only mode, but does not specify a 'report-to'; the policy will have no effect. Please either add a 'report-to' directive, or deliver the policy via the 'Content-Security-Policy' header.
CONSOLE MESSAGE: The Content Security Policy ' trusted-types tt-policy-name' was delivered in report-only mode, but does not specify a 'report-to'; the policy will have no effect. Please either add a 'report-to' directive, or deliver the policy via the 'Content-Security-Policy' header.
CONSOLE MESSAGE: Refused to create a TrustedTypePolicy named 'tt-policy-name' because it violates the following Content Security Policy directive: "trusted-types A"
CONSOLE MESSAGE: [Report Only] Refused to create a TrustedTypePolicy named 'tt-policy-name' because it violates the following Content Security Policy directive: "trusted-types B"
CONSOLE MESSAGE: [Report Only] Refused to create a TrustedTypePolicy named 'tt-policy-name' because it violates the following Content Security Policy directive: "trusted-types C"
CONSOLE MESSAGE: [Report Only] Refused to create a TrustedTypePolicy named 'tt-policy-name' because it violates the following Content Security Policy directive: "trusted-types D"
CONSOLE MESSAGE: [Report Only] Refused to create a TrustedTypePolicy named 'tt-policy-name' because it violates the following Content Security Policy directive: "trusted-types E"
CONSOLE MESSAGE: Refused to connect to ws://common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: The Content Security Policy 'trusted-types tt-policy-name' was delivered in report-only mode, but does not specify a 'report-to'; the policy will have no effect. Please either add a 'report-to' directive, or deliver the policy via the 'Content-Security-Policy' header.
CONSOLE MESSAGE: The Content Security Policy ' trusted-types other-policy-name' was delivered in report-only mode, but does not specify a 'report-to'; the policy will have no effect. Please either add a 'report-to' directive, or deliver the policy via the 'Content-Security-Policy' header.
CONSOLE MESSAGE: The Content Security Policy ' trusted-types *' was delivered in report-only mode, but does not specify a 'report-to'; the policy will have no effect. Please either add a 'report-to' directive, or deliver the policy via the 'Content-Security-Policy' header.
CONSOLE MESSAGE: The Content Security Policy ' trusted-types 'none'' was delivered in report-only mode, but does not specify a 'report-to'; the policy will have no effect. Please either add a 'report-to' directive, or deliver the policy via the 'Content-Security-Policy' header.
CONSOLE MESSAGE: The Content Security Policy ' trusted-types tt-policy-name 'allow-duplicates'' was delivered in report-only mode, but does not specify a 'report-to'; the policy will have no effect. Please either add a 'report-to' directive, or deliver the policy via the 'Content-Security-Policy' header.
CONSOLE MESSAGE: Refused to create a TrustedTypePolicy named 'tt-policy-name' because it violates the following Content Security Policy directive: "trusted-types other-policy-name"
CONSOLE MESSAGE: Refused to create a TrustedTypePolicy named 'tt-policy-name' because it violates the following Content Security Policy directive: "trusted-types 'none'"
CONSOLE MESSAGE: [Report Only] Refused to create a TrustedTypePolicy named 'tt-policy-name' because it violates the following Content Security Policy directive: "trusted-types other-policy-name"
CONSOLE MESSAGE: [Report Only] Refused to create a TrustedTypePolicy named 'tt-policy-name' because it violates the following Content Security Policy directive: "trusted-types 'none'"
CONSOLE MESSAGE: Refused to connect to ws://common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: The Content Security Policy 'trusted-types tt-policy-name' was delivered in report-only mode, but does not specify a 'report-to'; the policy will have no effect. Please either add a 'report-to' directive, or deliver the policy via the 'Content-Security-Policy' header.
CONSOLE MESSAGE: The Content Security Policy ' trusted-types *' was delivered in report-only mode, but does not specify a 'report-to'; the policy will have no effect. Please either add a 'report-to' directive, or deliver the policy via the 'Content-Security-Policy' header.
CONSOLE MESSAGE: The Content Security Policy ' trusted-types tt-policy-name 'allow-duplicates'' was delivered in report-only mode, but does not specify a 'report-to'; the policy will have no effect. Please either add a 'report-to' directive, or deliver the policy via the 'Content-Security-Policy' header.
CONSOLE MESSAGE: Refused to connect to ws://common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to create a TrustedTypePolicy named 'tt-policy-name' because it violates the following Content Security Policy directive: "trusted-types tt-policy-name"
CONSOLE MESSAGE: Refused to create a TrustedTypePolicy named 'tt-policy-name' because it violates the following Content Security Policy directive: "trusted-types *"
CONSOLE MESSAGE: [Report Only] Refused to create a TrustedTypePolicy named 'tt-policy-name' because it violates the following Content Security Policy directive: "trusted-types tt-policy-name"
CONSOLE MESSAGE: [Report Only] Refused to create a TrustedTypePolicy named 'tt-policy-name' because it violates the following Content Security Policy directive: "trusted-types *"
CONSOLE MESSAGE: Refused to connect to ws://common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.

PASS Multiple non-violated enforce trusted-types directives.
PASS Multiple report-only trusted-types directives.
PASS One violated report-only trusted-types directive followed by multiple enforce directives.
PASS One violated enforce trusted-types directive followed by multiple report-only directives.
PASS Mixing enforce and report-only policies with trusted-types directives
PASS Mixing enforce and report-only policies with trusted-types directives (duplicate policy)
